viewHistory finished. Only expDate left

+++ src/com/gafilianog/product/Cloth.java
@@ -5,7 +5,7 @@ public class Cloth extends Product {
 
     private final String size;
 
-    public Cloth(String name, double price, String size) {
+    public Cloth(String name, int price, String size) {
         super(name, price);
         this.size = size;
     }
