+++ src/com/gafilianog/product/Product.java
@@ -3,9 +3,9 @@ package com.gafilianog.product;
 public class Product {
 
     private String name;
-    private double price;
+    private int price;
 
-    public Product(String name, double price) {
+    public Product(String name, int price) {
         this.name = name;
         this.price = price;
     }
@@ -18,11 +18,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
 }
