+++ src/com/gafilianog/product/Tech.java
@@ -4,7 +4,7 @@ public class Tech extends Product {
 
     private final String version;
 
-    public Tech(String name, double price, String version) {
+    public Tech(String name, int price, String version) {
         super(name, price);
         this.version = version;
     }
